Add skydome shader (not wired up), extract NFS cars to own folders. Add HRZ parsing code for NFS3 track load.

+++ src/Loaders/trk_loader.h
@@ -46,7 +46,7 @@ public:
     std::vector<TrackBlock> track_blocks;
     std::vector<Entity> global_objects;
     uint32_t nBlocks;
-    map<short, GLuint> texture_gl_mappings;
+    map<unsigned int, GLuint> texture_gl_mappings;
 };
 
 class TrackLoader {
